Add files via upload
HR Data Analytics at Acme Telco

Acme Telco is a small telecom company that develops, sells, and installs telecommunication devices such as network routers, conference call hardware, etc. for business needs. Acme has been operating for over a decade with a modest current headcount of close to 200 people and it has organized itself around : Operations, IT, Sales, Software engineering, and Admin. And my role in this project is to give insights to the company to increase the growth of the company. Description of this project is to give insights about the company employee satisfaction. By using company data set, applying operations on Tableau